update table convert int number to float, update script to insert into monthlyPlan

+++ Services/SpendWiseAPI/Application/Interfaces/IParserData.cs
@@ -8,6 +8,6 @@ namespace Application.Interfaces
 {
     public interface IParserData
     {
-        string GetUpdatedStringSpentOfCategoryWhenAddTransaction(string categoryBudgetPlan, string spentOfCategoryMonthlyPlan, string CurrentCategory, int amount);
+        string GetUpdatedStringSpentOfCategoryWhenAddTransaction(string categoryBudgetPlan, string spentOfCategoryMonthlyPlan, string CurrentCategory, double amount);
     }
 }
